+++ src/Form/FormDtoDataTransformerInterface.php
@@ -2,12 +2,11 @@
 
 declare(strict_types=1);
 
-namespace App\Service\DataTransformer;
+namespace App\Form;
 
 use App\Entity\EntityInterface;
-use App\Form\FormDtoInterface;
 
-interface FormDtoTransformerInterface
+interface FormDtoDataTransformerInterface
 {
     /**
      * Convert Dto to new entity instance or fill existing instance entity with Dto data.
